+++ packages/freenodes/src/chains.js
@@ -3,6 +3,7 @@ export const CHAIN_BSC = 'bsc'
 export const CHAIN_XDAI = 'xdai'
 export const CHAIN_MATIC = 'matic'
 export const L2_ZKSYNC = 'zksync'
+export const L2_OPTIMISM = 'optimism'
 
 export const DEFAULT_CHAIN = CHAIN_ETHEREUM
 
@@ -12,5 +13,6 @@ export default {
   CHAIN_XDAI,
   CHAIN_MATIC,
   L2_ZKSYNC,
+  L2_OPTIMISM,
   DEFAULT_CHAIN,
 }