+++ packages/staker/src/bsc/utils.js
@@ -1,27 +1,11 @@
 'use strict';
 
-import { getRandomItem } from 'staker-freenodes';
+import { BSC_NODES, getRandomItem } from 'staker-freenodes';
 
 import {getConfig} from '../utils';
 import {jsonRpcFetch} from '../helpers/jsonRpc';
 
 // ==== Binance Smart Chain JSON RPC ===
-
-
-export const BSC_NODES = [
-  'https://bsc-dataseed1.binance.org',
-  'https://bsc-dataseed2.binance.org',
-  'https://bsc-dataseed3.binance.org',
-  'https://bsc-dataseed4.binance.org',
-  'https://bsc-dataseed1.defibit.io',
-  'https://bsc-dataseed2.defibit.io',
-  'https://bsc-dataseed3.defibit.io',
-  'https://bsc-dataseed4.defibit.io',
-  'https://bsc-dataseed1.ninicoin.io',
-  'https://bsc-dataseed2.ninicoin.io',
-  'https://bsc-dataseed3.ninicoin.io',
-  'https://bsc-dataseed4.ninicoin.io',
-];
 
 let cachedBscNodeUrl = '';