+++ packages/staker/src/cli.js
@@ -32,6 +32,7 @@ import {skills as skillsXDAI} from './xdai/saihubot-cli-skill-xdai';
 import {skills as skillsBSC} from './bsc/saihubot-cli-skill-bsc';
 import {skills as skillsMatic} from './matic/saihubot-cli-skill-matic';
 import {skills as skillsZkSync} from './zksync/saihubot-cli-skill-zksync';
+import {skills as skillsOptimism} from './zksync/saihubot-cli-skill-optimism';
 import {skills as skillsChain} from './saihubot-cli-skill-chainid';
 import {skillAwesomeEth} from './saihubot-skill-awesome';
 
@@ -83,6 +84,7 @@ const bot = new SaihuBot({
     ...skillsMatic,
     ...skillsXDAI,
     ...skillsZkSync,
+    ...skillsOptimism,
     skillAwesomeEth,
     ...skillDelegates,
     skillHelp,